requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-environ>=0.9
django-filter>=23.0
